backend/requirements.txt
anthropic>=0.40.0
google-generativeai>=0.4.0
openai>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pymupdf>=1.23.0
pillow>=10.2.0
python-dotenv>=1.0.0
numpy>=1.26.0
aiofiles>=23.2.1
httpx>=0.26.0
